replace writeCitoids
